Extract shared base modal Stimulus controller
The cookie-consent, location and time controllers each reimplemented
the same connect/disconnect/close/closeOnEscape modal logic, differing
only by a hardcoded turbo-frame id. Add a base Modal controller that
closes by clearing its enclosing turbo-frame (via closest), removing
the id coupling, and have the three controllers extend it.
