+++ app/javascript/controllers/cookie_consent_modal_controller.js
@@ -1,25 +1,3 @@
-import { Controller } from '@hotwired/stimulus'
+import Modal from 'controllers/modal'
 
-export default class extends Controller {
-  connect () {
-    this.boundCloseOnEscape = this.closeOnEscape.bind(this)
-    document.addEventListener('keydown', this.boundCloseOnEscape)
-  }
-
-  disconnect () {
-    document.removeEventListener('keydown', this.boundCloseOnEscape)
-  }
-
-  close () {
-    const modalFrame = document.getElementById('cookie_consent_modal')
-    if (modalFrame) {
-      modalFrame.innerHTML = ''
-    }
-  }
-
-  closeOnEscape (event) {
-    if (event.key === 'Escape') {
-      this.close()
-    }
-  }
-}
+export default class extends Modal {}
